LangChain_RAG_PDF_Uploads_Chat_History/requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-community>=0.2.9
langchain-groq>=0.2.0
chromadb>=1.0.0
sentence-transformers>=0.2.0
pypdf>=5.0.0